remapping
-changed some ui elements
-moved checklistbutton to slideshow page
