+++ app/src/main/java/com/kip/receiptscanner/MainDrawer.kt
@@ -42,11 +42,14 @@ class MainDrawer : AppCompatActivity() {
         navView.setupWithNavController(navController)
     }
 
+    //disable/enable menu button
+    /*
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
         // Inflate the menu; this adds items to the action bar if it is present.
         menuInflater.inflate(R.menu.main_drawer, menu)
         return true
     }
+     */
 
     override fun onSupportNavigateUp(): Boolean {
         val navController = findNavController(R.id.nav_host_fragment)
