+++ app/src/main/java/com/kip/receiptscanner/ui/home/HomeFragment.kt
@@ -33,18 +33,4 @@ class HomeFragment : Fragment() {
         return root
     }
 
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        button.setOnClickListener {
-            var intent = Intent(
-                view.context,
-                ChecklistActivity::class.java
-            )
-            startActivity(intent)
-        }
-
-
-
-    }
 }
