+++ app/src/main/java/com/kip/receiptscanner/ui/slideshow/SlideshowFragment.kt
@@ -1,5 +1,6 @@
 package com.kip.receiptscanner.ui.slideshow
 
+import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -9,6 +10,8 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
 import com.kip.receiptscanner.R
+import com.kip.receiptscanner.checklist.ChecklistActivity
+import kotlinx.android.synthetic.main.fragment_slideshow.*
 
 class SlideshowFragment : Fragment() {
 
@@ -27,5 +30,17 @@ class SlideshowFragment : Fragment() {
             textView.text = it
         })
         return root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        button.setOnClickListener {
+            var intent = Intent(
+                view.context,
+                ChecklistActivity::class.java
+            )
+            startActivity(intent)
+        }
     }
 }
